GlobalTradeAnalysisProjectModelV7

This is located in docs/src/index.md. All of this is in Markdown. I haven't done anything too fancy with your documentation, but I did add docstrings to all exported functions and types, which is required.

You can also show example code:

using Pkg

Pkg.add("GlobalTradeAnalysisProjectModelV7")

You can even run code in documentation:

x = 3

x^2

9

GlobalTradeAnalysisProjectModelV7.aggregate_data — Method

    aggregate_data(; hData, hParameters, hSets, comMap, regMap, endMap)

YOu can type math here $\int_0^1\,x^2\,dx$. And you can do fancy math

\[e^{i\pi} + 1 = 0\]

This is all in markdown. You can also use latex in the code blocks.

GlobalTradeAnalysisProjectModelV7.aggregate_data_legacy — Method

    aggregate_data_legacy(; hData, hParameters, hSets, comMap, regMap, endMap)

GlobalTradeAnalysisProjectModelV7.build_model! — Method

build_model!(mc; max_iter=50, constr_viol_tol=1e-8, bound_push=1e-15)

GlobalTradeAnalysisProjectModelV7.calculate_ev — Method

calculate_ev(; sets, data0, data1, parameters, max_iter=50, constr_viol_tol=1e-5, bound_push=1e-15)

Doc string

GlobalTradeAnalysisProjectModelV7.calculate_gdp — Method

calculate_gdp(; sets, data0, data1)

GlobalTradeAnalysisProjectModelV7.extract_model! — Method

extract_model!(model_container)

GlobalTradeAnalysisProjectModelV7.generate_calibration_inputs — Method

generate_calibration_inputs(model_container, start_data)

GlobalTradeAnalysisProjectModelV7.generate_initial_model — Method

generate_initial_model(; hSets, hData, hParameters)

GlobalTradeAnalysisProjectModelV7.generate_starting_values — Method

generate_starting_values(; hSets, hData, hParameters)

Doc string here. It generates sets.

GlobalTradeAnalysisProjectModelV7.get_sample_data — Method

get_sample_data()

GlobalTradeAnalysisProjectModelV7.initialize_model! — Method

initialize_model!(model_container)

GlobalTradeAnalysisProjectModelV7.run_model! — Method

run_model!(; model_container, max_iter=50, constr_viol_tol=1e-8, bound_push=1e-15)

Runs the model (modelcontainer.model) by executing these steps: (1) set all starting values to those found in the data (modelcontainer.data), (2) fix all variables that are specified as exogenous (modelcontainer.fixed), (3) solve the model, (4) load all variable values to the data (modelcontainer.data)

GlobalTradeAnalysisProjectModelV7.solve_model! — Method

solve_model!(model_container; max_iter, constr_viol_tol, bound_push)

GlobalTradeAnalysisProjectModelV7.solve_model — Method

solve_model(; sets, data, parameters, fixed, max_iter=50, constr_viol_tol=1e-8, bound_push=1e-15, model=nothing)

GlobalTradeAnalysisProjectModelV7.model_container_struct — Type

mutable struct model_container_struct

A container of all data for a model